cais6/outputs/code/optimizerphasetransitions/requirements.txt
numpy
scipy
scikit-learn
torch
torchvision
matplotlib
pyyaml
tqdm
statsmodels